Check for missing voice data only when we get error in speak
Oddly, TTS voice.getFeature returns notinstalled sometimes even when the data is installed. Because of this false positive we were showing a dialog to install voice data when it was already installed. With this change we now move the voice data check outside, and only report it when we get an error while speaking.