Handle search text for orientation changes
On orientation change, searchView is null until onOptionsMenuItem is called. But if we're getting a barcode result at the same time then onActivityResult will be called before onOptionsMenuItem thus the call to setSearchText fails because searchview is not set at that time.
This change will save the text in a variable and then apply the same when onOptionsMenuItem is triggered.